# swagger.yaml
openapi: 3.0.3
info:
  title: TraintiQ Profile Generator API
  description: |
    # 🚀 Advanced Company Profile Generation System
    
    ## Overview
    Comprehensive API for AI-powered company profile generation using multi-source data extraction and advanced natural language processing.
    
    ## 🌟 Key Features
    - **AI-Powered Generation**: OpenAI GPT-4 for intelligent content synthesis
    - **Multi-Source Analysis**: Web scraping, document processing, social media integration
    - **Enhanced NLP/NER**: Advanced entity extraction and content categorization
    - **Real-time Processing**: Asynchronous generation with progress tracking
    - **Template System**: Industry-specific profile templates
    - **Quality Assessment**: Source analysis and content validation
    
    ## 🔧 Technology Stack
    - **Backend**: Flask + FastAPI hybrid architecture
    - **AI/ML**: OpenAI GPT-4, LangChain, Sentence Transformers
    - **Data Processing**: BeautifulSoup, Selenium, PyPDF2, spaCy
    - **Database**: MySQL with SQLAlchemy ORM
    - **Async Processing**: Celery with Redis
    - **Monitoring**: Prometheus metrics and health checks
    
    ## 🚦 Rate Limits & Performance
    - **Profile Generation**: 10 requests/minute (avg. 10-15 seconds per request)
    - **Source Analysis**: 20 requests/minute (avg. 3-5 seconds per request)
    - **Template Operations**: 50 requests/minute (instant response)
    - **Success Rate**: 95%+ for valid URLs
...
    
  version: 2.0.0
  contact:
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
  /profile/generate:
    post:
      tags:
        - Profile Generation
      summary: Generate Enhanced Company Profile
      description: |
        Generate a comprehensive company profile using AI-powered analysis of multiple data sources.
        
        ## Process Overview:
        1. **Source Validation**: Validate and analyze provided URLs
        2. **Data Extraction**: Scrape content using advanced techniques (requests + Selenium fallback)
        3. **AI Processing**: Use OpenAI GPT-4 for intelligent content synthesis
        4. **Structured Output**: Organize data into categorized sections
        5. **Quality Assessment**: Validate and score generated content
        
        ## Supported Data Sources:
        - Company websites and landing pages
        - LinkedIn company profiles
        - Crunchbase and similar business directories
        - Social media profiles
        - News articles and press releases
      operationId: generateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileGenerationRequest'
            examples:
              technology_company:
                summary: Technology Company Profile
                value:
                  urls:
                    - "https://www.softcodeit.com"
                    - "https://linkedin.com/company/softcodeit"
                  template: "technology"
                  focus_areas: ["products", "technology", "leadership"]
                  custom_text: "Focus on AI and machine learning capabilities"
              startup_company:
                summary: Startup Company Profile
                value:
                  urls:
                    - "https://www.startup.com"
                  template: "startup"
                  focus_areas: ["overview", "products", "funding"]
                  include_financial: false
      responses:
        '200':
          description: Profile generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileGenerationResponse'
              examples:
                successful_generation:
                  summary: Successful Profile Generation
                  value:
                    success: true
                    profile_data:
                      basic_info:
                        company_name: "SoftCodeIT Solutions"
                        company_overview: "Leading software development company"
                        industry: "Software Development"
                        founded_year: "2018"
                        headquarters: "Colombo, Sri Lanka"
                        company_size: "50-100 employees"
                        website: "https://www.softcodeit.com"
                      products_services:
                        - name: "Custom Web Development"
                          description: "End-to-end web application development"
                          category: "Web Development"
                          features: ["React/Angular", "Node.js/Python", "Cloud deployment"]
                      leadership_team:
                        - name: "John Smith"
                          position: "Chief Technology Officer"
                          bio: "15+ years experience in software architecture"
                          experience: "15+ years"
                      technology_stack:
                        - name: "React"
                          category: "Frontend Framework"
                          description: "Primary frontend framework"
                          version: "18.x"
                      contact_info:
                        email: "info@softcodeit.com"
                        phone: "+94 11 234 5678"
                        address: "123 Tech Street, Colombo 03, Sri Lanka"
                      company_values: ["Innovation", "Quality", "Customer Focus"]
                      achievements: ["ISO 9001 Certified", "Top Software Company 2023"]
                    metadata:
                      generation_time: 12.45
                      sources_processed: 2
                      ai_model_used: "gpt-4-turbo"
                      content_quality_score: 0.87
                      tokens_used: 3420
                    message: "Enhanced profile generated successfully"
        '400':
          description: Bad Request - Invalid input parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error - Generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile/analyze/sources:
    get:
      tags:
        - Source Analysis
      summary: Analyze Data Sources Quality
      description: |
        Perform comprehensive analysis of data sources to assess their suitability for profile generation.
        
        ## Analysis Capabilities:
        - **Content Quality**: Assess content richness and relevance
        - **Accessibility**: Test URL accessibility and response times
        - **Data Extraction**: Validate successful content extraction
        - **Company Information**: Detect presence of company-specific data
        - **Performance Metrics**: Measure response times and content volume
      operationId: analyzeDataSources
      parameters:
        - name: urls
          in: query
          required: true
          description: URLs to analyze (can be repeated)
          schema:
            type: array
            items:
              type: string
              format: uri
          style: form
          explode: true
          examples:
            single_url:
              summary: Single URL Analysis
              value: ["https://www.softcodeit.com"]
            multiple_urls:
              summary: Multiple URLs Analysis
              value: ["https://www.softcodeit.com", "https://linkedin.com/company/softcodeit"]
      responses:
        '200':
          description: Source analysis completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceAnalysisResponse'
        '400':
          description: Bad Request - No URLs provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error - Analysis failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile/templates:
    get:
      tags:
        - Template Management
      summary: Get Available Profile Templates
      description: |
        Retrieve all available profile templates with their configurations and focus areas.
        
        ## Available Templates:
        - **startup**: Perfect for early-stage companies and startups
        - **enterprise**: Comprehensive profile for established enterprises
        - **technology**: Specialized for tech companies and software businesses
        - **financial**: Tailored for banks, fintech, and financial institutions
      operationId: getProfileTemplates
      responses:
        '200':
          description: Templates retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplatesResponse'
        '500':
          description: Internal Server Error - Template retrieval failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for authentication (production only)
  schemas:
    ProfileGenerationRequest:
      type: object
      required:
        - urls
      properties:
        urls:
          type: array
          items:
            type: string
            format: uri
          description: URLs to analyze for profile generation
          example: ["https://www.softcodeit.com", "https://linkedin.com/company/softcodeit"]
          minItems: 1
          maxItems: 10
        custom_text:
          type: string
          description: Additional context or instructions
          example: "Focus on the company's AI and machine learning capabilities"
          maxLength: 2000
        focus_areas:
          type: array
          items:
            type: string
            enum: [overview, products, leadership, technology, market, funding, history, financials, compliance, innovation, achievements]
          description: Specific areas to emphasize
          example: ["products", "leadership", "technology", "achievements"]
        template:
          type: string
          enum: [startup, enterprise, technology, financial]
          description: Profile template to use
          example: "technology"
        include_financial:
          type: boolean
          default: false
          description: Include financial information if available
        include_news:
          type: boolean
          default: true
          description: Include recent news and updates
        max_content_length:
          type: integer
          default: 5000
          minimum: 1000
          maximum: 10000
          description: Maximum content length per section
        language:
          type: string
          default: "en"
          enum: [en, es, fr, de, it, pt, ja, ko, zh]
          description: Content language
...
    ProfileData:
      type: object
      required:
        - basic_info
      properties:
        basic_info:
          $ref: '#/components/schemas/CompanyBasicInfo'
        products_services:
          type: array
          items:
            $ref: '#/components/schemas/ProductService'
          description: Products and services portfolio
        leadership_team:
          type: array
          items:
            $ref: '#/components/schemas/LeadershipMember'
          description: Key leadership members
        technology_stack:
          type: array
          items:
            $ref: '#/components/schemas/TechnologyItem'
          description: Technology stack and tools
        contact_info:
          $ref: '#/components/schemas/ContactInfo'
        company_values:
          type: array
          items:
            type: string
          description: Core values and principles
          example: ["Innovation", "Quality", "Customer Focus", "Integrity"]
        achievements:
          type: array
          items:
            type: string
          description: Notable achievements
          example: ["ISO 9001 Certified", "Top Software Company 2023", "500+ Projects Delivered"]
        market_position:
          type: string
          description: Market positioning
          example: "Leading provider of custom software solutions in South Asia"
        recent_news:
          type: array
          items:
            type: string
          description: Recent updates and news
        financial_info:
          type: object
          description: Financial information (if available)
    GenerationMetadata:
      type: object
      required:
        - generation_time
        - sources_processed
        - ai_model_used
        - generation_timestamp
      properties:
        generation_time:
          type: number
          format: float
          description: Generation time in seconds
          example: 12.45
        sources_processed:
          type: integer
          description: Successfully processed sources
          example: 2
        ai_model_used:
          type: string
          description: AI model identifier
          example: "gpt-4-turbo"
        content_quality_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Content quality score (0-1)
          example: 0.87
        data_sources:
          type: array
          items:
            type: string
            format: uri
          description: Source URLs used
          example: ["https://www.softcodeit.com", "https://linkedin.com/company/softcodeit"]
        processing_method:
          type: string
          description: Processing approach
          example: "enhanced_ai_generation"
        tokens_used:
          type: integer
          description: AI tokens consumed
          example: 3420
        generation_timestamp:
          type: string
          format: date-time
          description: Generation completion time
    ProfileGenerationResponse:
      type: object
      required:
        - success
        - profile_data
        - metadata
        - message
      properties:
        success:
          type: boolean
          example: true
        profile_data:
          $ref: '#/components/schemas/ProfileData'
        metadata:
          $ref: '#/components/schemas/GenerationMetadata'
        message:
          type: string
          example: "Enhanced profile generated successfully"
...
    SourceAnalysisSummary:
      type: object
      required:
        - total_sources
        - valid_sources
        - successful_scrapes
        - failed_scrapes
        - total_content_length
        - average_content_length
        - source_quality
        - analysis_timestamp
      properties:
        total_sources:
          type: integer
          description: Total URLs provided
          example: 2
        valid_sources:
          type: integer
          description: Valid HTTP/HTTPS URLs
          example: 2
        successful_scrapes:
          type: integer
          description: Successfully analyzed URLs
          example: 2
        failed_scrapes:
          type: integer
          description: Failed analyses
          example: 0
        total_content_length:
          type: integer
          description: Total content extracted (characters)
          example: 45680
        average_content_length:
          type: number
          format: float
          description: Average content per source
          example: 22840.0
        source_quality:
          type: string
          enum: [excellent, good, poor]
          description: Overall quality assessment
          example: "excellent"
        domains_analyzed:
          type: array
          items:
            type: string
          description: Analyzed domains
          example: ["softcodeit.com", "linkedin.com"]
        detailed_results:
          type: array
          items:
            $ref: '#/components/schemas/SourceAnalysisDetail'
          description: Per-URL analysis results
        analysis_timestamp:
          type: string
          format: date-time
          description: Analysis completion time
    SourceAnalysisResponse:
      type: object
      required:
        - success
        - analysis
        - recommendations
        - message
      properties:
        success:
          type: boolean
          example: true
        analysis:
          $ref: '#/components/schemas/SourceAnalysisSummary'
        recommendations:
          type: array
          items:
            type: string
          description: Actionable recommendations
          example: ["✅ All URLs successfully analyzed", "📄 Rich content detected - excellent for comprehensive profile"]
        message:
          type: string
          example: "Real source analysis completed successfully"
    ProfileTemplate:
      type: object
      required:
        - name
        - description
        - focus_areas
      properties:
        name:
          type: string
          description: Template display name
          example: "Technology Company"
        description:
          type: string
          description: Template purpose and use case
          example: "Specialized for tech companies and software businesses"
        focus_areas:
          type: array
          items:
            type: string
          description: Key areas this template emphasizes
          example: ["overview", "products", "technology", "leadership", "market", "innovation"]
    TemplatesResponse:
      type: object
      required:
        - success
        - templates
        - message
      properties:
        success:
          type: boolean
          example: true
        templates:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ProfileTemplate'
          description: Dictionary of available profile templates
        message:
          type: string
          example: "Templates retrieved successfully"
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - message
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error type or classification
          example: "ValidationError"
        message:
          type: string
          description: Detailed error description
          example: "Invalid URL format provided"
        details:
          type: object
          description: Additional error context and debugging information
        timestamp:
          type: string
          format: date-time
          description: Error occurrence timestamp
...